drv_r Creates a Nix expression running an R function

drv_r Creates a Nix expression running an R function

Usage

drv_r(name, expr)

Arguments

name

Symbol, name of the derivation.

expr

R code to generate the expression.

output

Character, a variable name to save the output of the function expression

Value

A list of two elements, name, the name of the derivation, and snippet the Nix boilerplate code.

Details

At a basic level, drv_r(mtcars_am, filter(mtcars, am == 1)) is equivalent to mtcars <- filter(mtcars, am == 1). drv_r() generates the required Nix boilerplate to output a so-called "derivation" in Nix jargon. A Nix derivation is a recipe that defines how to create an output (in this case mtcars_am) including its dependencies, build steps, and output paths.

Examples

drv_r(mtcars_am, filter(mtcars, am == 1))
#> $name
#> [1] "mtcars_am"
#> 
#> $snippet
#> [1] "  mtcars_am = makeRDerivation {\n    name = \"mtcars_am\";\n    buildPhase = ''\n      Rscript -e \"\n        source('libraries.R')\n        mtcars_am <- filter(mtcars, am == 1)\n        saveRDS(mtcars_am, 'mtcars_am.rds')\"\n    '';\n  };"
#>