Generate Nix Pipeline Code with a Generic Default Target

This function generates a string containing Nix code for a pipeline.nix file based on a list of derivation objects. Each derivation is defined by a snippet of Nix code and a name, typically created using a function like rxp_r. The resulting Nix code defines all derivations and sets a generic default target that builds all of them using pkgs.symlinkJoin. This is designed for building reproducible analytical pipelines with Nix, inspired by the targets R package.

Usage

rixpress(derivs, project_path)

Arguments

derivs

A list of derivation objects, where each object must have two elements: $name (a character string naming the derivation) and $snippet (a character string containing the Nix code defining the derivation). Typically, these objects are created by a function like rxp_r.

project_path

Path to root of project, typically "."

Value

A character string containing the complete Nix code for a pipeline.nix file. This string can be written to a file or passed to another function for further processing.

Details

The generated Nix code includes:

A default.nix import for environment setup via the rix package.

A makeRDerivation function to standardize derivation creation.

All derivations from the derivs list.

A default target (allDerivations) that builds all derivations when nix-build pipeline.nix is run without specifying an attribute.

The function assumes the existence of a libraries.R file in the working directory, which is copied into the build environment by the Nix code.

Examples

if (FALSE) { # \dontrun{
# Create derivation objects
d1 <- rxp_r(mtcars_am, filter(mtcars, am == 1))
d2 <- rxp_r(mtcars_head, head(mtcars_am))
list_derivs <- list(d1, d2)

# Generate the pipeline code
rixpress(derivs = list_derivs, project_path = ".")

} # }