Run a pipeline on GitHub Actions

Run a pipeline on GitHub Actions

Usage

rxp_ga()

Value

Nothing, copies file to a directory.

Details

This function puts a .yaml file inside the .github/workflows/ folder on the root of your project. This workflow file expects both scripts generated by rxp_init(), gen-env.R and gen-pipeline.R to be present. If that's not the case, edit the .yaml file accordingly. Build artifacts are archived and restored automatically between runs. Make sure to give read and write permissions to the GitHub Actions bot.

See also

Examples

if (FALSE) { # \dontrun{
  rxp_ga()
} # }