Create a Nix derivation for Python code execution

This function generates a Nix derivation that executes a given Python expression and serializes the result using Python's pickle module. The derivation is designed to work with Python environments defined in a default.nix file.

Usage

rxp_py(name, py_expr, nix_env = "default.nix")

Arguments

name

A symbol representing the name of the variable to be serialized and the derivation name. It must be a valid Python variable name (e.g., no dots).

py_expr

A string containing the Python expression to be assigned to the variable specified by name. The function will execute <name> = <py_expr> in Python.

nix_env

The path to the Nix environment file (default: "default.nix"). This file should define the Python packages required for the derivation.

Value

A list containing the derivation name, the Nix derivation snippet, the type of derivation ("rxp_py"), and the Nix environment setup code.

Examples

# Generate a derivation that assigns 42 to 'my_result' and pickles it
rxp_py(my_result, "42")
#> $name
#> [1] "my_result"
#> 
#> $snippet
#> [1] "  my_result = makePyDerivation {\n    name = \"my_result\";\n    buildInputs = defaultBuildInputs;\n    configurePhase = defaultConfigurePhase;\n    buildPhase = ''\n      python -c \"\nexec(open('libraries.py').read())\nexec('my_result = 42')\nwith open('my_result.pickle', 'wb') as f: pickle.dump(globals()['my_result'], f)\"\n    '';\n  };"
#> 
#> $type
#> [1] "rxp_py"
#> 
#> $nix_env
#> [1] "default = import ./default.nix;\n  defaultPkgs = default.pkgs;\n  defaultShell = default.shell;\n  defaultBuildInputs = defaultShell.buildInputs;\n  defaultConfigurePhase = ''\n    cp ${./_rixpress/default_libraries.py} libraries.py\n    mkdir -p $out\n  '';"
#> 
# The generated Python code will be:
# exec(open('libraries.py').read()); exec('my_result = 42');
# import pickle; with open('my_result.pickle', 'wb') as f:
#     pickle.dump(globals()['my_result'], f)