Read output of a derivation

Reads the output of derivations in the current session, returns a path if reading directly is not possible.

Usage

rxp_read(derivation_name, which_log = NULL, project_path = ".")

Arguments

derivation_name

Character, the name of the derivation.

which_log

Character, defaults to NULL. If NULL the most recent build log is used. If a string is provided, it's used as a regular expression to match against available log files.

project_path

Character, defaults to ".". Path to the root directory of the project.

Value

The derivation's output.

Details

When derivation_name points to a single R object, it gets read in the current session using readRDS(). If it's a Python object and {reticulate} is available, reticulate::py_load_object() is used. In case the derivation is pointing to several outputs (which can happen when building a Quarto document for example) or neither readRDS() nor reticulate::py_load_object() successfully read the object, the path to the object is returned instead.

See also

Examples

if (FALSE) { # \dontrun{
  mtcars <- rxp_read("mtcars")

  # Read from a specific build log
  mtcars <- rxp_read("mtcars", which_log = "2025-05-10")
} # }